Refactored ApiModal subcomponents into separate components, rm redux-dev
refactored apimodal subcomponents fullcontent, preview content, and minimized content into separate components. minor revisions to props for each to allow management of css properties via api. 

removed redux-devtools package from package.json and packagelock, was for use in local dev env only.
