+++ package.json
@@ -98,7 +98,6 @@
     ]
   },
   "devDependencies": {
-    "node-sass": "^4.14.1",
-    "redux-devtools-extension": "^2.13.8"
+    "node-sass": "^4.14.1"
   }
 }
