+++ src/actions/apiModalActions.js
@@ -47,7 +47,7 @@ export function fetchModals() {
     {
       modal_id: "21345",
       default_content_state: "preview",
-      display_delay: 10,
+      display_delay: 5,
       preview_text: "Preview Text 1",
       preview_header: "Preview Header 1",
       accept_button_text: "Accept",
@@ -60,6 +60,7 @@ export function fetchModals() {
       full_header: "Full Header 1",
       full_position: null,
       full_size: null,
+      full_background_color: null,
       survey_id: "dd842d3c-04b0-4842-b076-63446439b3a6",
       sheet_id: null,
       minimized_icon: "maximize",
@@ -68,20 +69,21 @@ export function fetchModals() {
     {
       modal_id: "2134sdf5",
       default_content_state: "preview",
-      display_delay: 360,
+      display_delay: 10,
       preview_text: "Preview Text 2",
       preview_header: "Preview Header 2",
       accept_button_text: "Accept",
       later_button_text: "Maybe Later",
       reject_button_text: "No Thanks",
-      preview_position: null,
+      preview_position: 'center',
       preview_size: null,
       preview_background_color: null,
       full_text: "Full Text 2",
       full_header: "Full Header 2",
-      full_position: null,
-      full_size: null,
-      survey_id: "dd842d3c-04b0-4842-b076-63446439b3a6",
+      full_position: 'center',
+      full_size: 'full-screen',
+      full_background_color: '#ccccccc5',
+      survey_id: null,
       sheet_id: null,
       minimized_icon: "maximize",
       minimized_text: "TNRIS Survey",
